Make set_ and clear_buffer_xy ignore out of range values

+++ firmware/display.h
@@ -45,7 +45,7 @@ void display_set_day(bool day);
 
 void display_show_buffer(void);
 void display_setbuffer_xy(int x, int y);
-void display_clearbuffer_xy(uint8_t x, uint8_t y);
+void display_clearbuffer_xy(int x, int y);
 void display_draw_line(int x0, int y0, int x1, int y1);
 #endif
 #endif
